fix: remove cycle dependencies
