+++ aemedge/blocks/cards/cards.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import { readBlockConfig } from '../../scripts/aem.js';
-import { createOptimizedPicture } from '../../scripts/scripts.js';
+import createOptimizedPicture from '../../scripts/utils/picture.js';
 import { fetchAndFilterDataIndex } from '../../scripts/indexing.js';
 import {
   createElement,
