+++ aemedge/scripts/utils/picture.js
@@ -1,0 +1,53 @@
+import { createOptimizedPicture as libCreateOptimizedPicture } from '../aem.js';
+import { appendQueryParams } from './uri.js';
+
+/**
+ * Creates an optimized picture element for an image.
+ * If the image is not an absolute URL, it will be passed to libCreateOptimizedPicture.
+ * @param {string} src The image source URL
+ * @param {string} alt The image alt text
+ * @param {boolean} eager Whether to load the image eagerly
+ * @param {object[]} breakpoints The breakpoints to use
+ * @returns {Element} The picture element
+ */
+export default function createOptimizedPicture(src, alt = '', eager = false, breakpoints = [{ media: '(min-width: 600px)', width: '2000' }, { width: '750' }]) {
+  const isAbsoluteUrl = /^https?:\/\//i.test(src);
+
+  // Fallback to createOptimizedPicture if src is not an absolute URL
+  if (!isAbsoluteUrl) return libCreateOptimizedPicture(src, alt, eager, breakpoints);
+
+  const url = new URL(src);
+  const picture = document.createElement('picture');
+  const { pathname } = url;
+  const ext = pathname.substring(pathname.lastIndexOf('.') + 1);
+
+  // webp
+  breakpoints.forEach((br) => {
+    const source = document.createElement('source');
+    if (br.media) source.setAttribute('media', br.media);
+    source.setAttribute('type', 'image/webp');
+    const searchParams = new URLSearchParams({ width: br.width, format: 'webply' });
+    source.setAttribute('srcset', appendQueryParams(url, searchParams));
+    picture.appendChild(source);
+  });
+
+  // fallback
+  breakpoints.forEach((br, i) => {
+    const searchParams = new URLSearchParams({ width: br.width, format: ext });
+
+    if (i < breakpoints.length - 1) {
+      const source = document.createElement('source');
+      if (br.media) source.setAttribute('media', br.media);
+      source.setAttribute('srcset', appendQueryParams(url, searchParams));
+      picture.appendChild(source);
+    } else {
+      const img = document.createElement('img');
+      img.setAttribute('loading', eager ? 'eager' : 'lazy');
+      img.setAttribute('alt', alt);
+      picture.appendChild(img);
+      img.setAttribute('src', appendQueryParams(url, searchParams));
+    }
+  });
+
+  return picture;
+}
